+++ core/src/geometry/size.rs
@@ -157,7 +157,7 @@ impl Size {
     /// Saturating addition.
     ///
     /// Returns `u32::max_value()` for `width` and/or `height` instead of overflowing.
-    pub fn saturating_add(self, other: Self) -> Self {
+    pub const fn saturating_add(self, other: Self) -> Self {
         Self {
             width: self.width.saturating_add(other.width),
             height: self.height.saturating_add(other.height),
@@ -168,11 +168,18 @@ impl Size {
     ///
     /// Returns `0` for `width` and/or `height` instead of overflowing, if the
     /// value in `other` is larger then in `self`.
-    pub fn saturating_sub(self, other: Self) -> Self {
+    pub const fn saturating_sub(self, other: Self) -> Self {
         Self {
             width: self.width.saturating_sub(other.width),
             height: self.height.saturating_sub(other.height),
         }
+    }
+
+    /// Division.
+    ///
+    /// This method provides a workaround for the `Div` trait not being usable in `const` contexts.
+    pub(crate) const fn div_u32(self, rhs: u32) -> Size {
+        Size::new(self.width / rhs, self.height / rhs)
     }
 
     /// Creates a size from two corner points of a bounding box.
@@ -218,7 +225,7 @@ impl Size {
     ///
     /// assert_eq!(result, Size::new(40, 90));
     /// ```
-    pub fn component_mul(self, other: Self) -> Self {
+    pub const fn component_mul(self, other: Self) -> Self {
         Self::new(self.width * other.width, self.height * other.height)
     }
 
@@ -235,7 +242,7 @@ impl Size {
     ///
     /// assert_eq!(result, Size::new(4, 3));
     /// ```
-    pub fn component_div(self, other: Self) -> Self {
+    pub const fn component_div(self, other: Self) -> Self {
         Self::new(self.width / other.width, self.height / other.height)
     }
 }
@@ -289,7 +296,7 @@ impl Div<u32> for Size {
     type Output = Size;
 
     fn div(self, rhs: u32) -> Size {
-        Size::new(self.width / rhs, self.height / rhs)
+        self.div_u32(rhs)
     }
 }