+++ core/src/pixelcolor/binary_color.rs
@@ -66,7 +66,7 @@ impl BinaryColor {
     /// assert_eq!(BinaryColor::On.invert(), BinaryColor::Off);
     /// ```
     #[inline]
-    pub fn invert(self) -> Self {
+    pub const fn invert(self) -> Self {
         match self {
             BinaryColor::On => BinaryColor::Off,
             BinaryColor::Off => BinaryColor::On,
@@ -83,8 +83,8 @@ impl BinaryColor {
     /// assert!(BinaryColor::On.is_on());
     /// ```
     #[inline]
-    pub fn is_on(self) -> bool {
-        self == BinaryColor::On
+    pub const fn is_on(self) -> bool {
+        matches!(self, BinaryColor::On)
     }
 
     /// Returns `true` if this color is `Off`.
@@ -97,8 +97,8 @@ impl BinaryColor {
     /// assert!(BinaryColor::Off.is_off());
     /// ```
     #[inline]
-    pub fn is_off(self) -> bool {
-        self == BinaryColor::Off
+    pub const fn is_off(self) -> bool {
+        matches!(self, BinaryColor::Off)
     }
 
     /// Maps active and inactive colors to a different type.